no borrar snapshot si no se hace rolback. Ajuste apariencia backup html

+++ app/lib/backup.rb
@@ -219,7 +219,7 @@ module Backup
       raise e
     ensure
       FileUtils.rm_rf(restore_dir)
-      FileUtils.rm_rf(snapshot_path)
+      FileUtils.rm_rf(snapshot_path) if rollback
     end
   end
